fix Node destruction on virtual method calls.
- Fix Node destroying on call of any first virtual method of its class,
caused by incorrect mixture of virtual methods with preprocessor
directives in node's header (messed with vtable and caused game
compiled without DRAW_DEBUG to have one less virtual method in table
and thus always execute the previous one for Node. Which happend to be
a destructor). Huge thanks to @alex-eg for figuring this out!