rework node alignment (again)
- Add virtual Vector2 Node.get_offset(Align). Name may be changed, but
the whole point is to return offset pos of the specific node's side.
May need to cache it later.
- Rework node alignment logic - once again is uses current's node
alignment to affect it (instead of affecting the ancestor), but this
time it also calculates parent's offset. Seems to be the better way
to do that. May need to allow setting these two separately later